Google sync fixes and inactive contact handling

+++ app/models/google_contacts_integrator.rb
@@ -6,9 +6,20 @@ class GoogleContactsIntegrator
 
   CONTACTS_GROUP_TITLE = 'MPDX'
 
+  # If a contact in MPDX gets marked as inactive, e.g. 'Not Interested', then they won't be synced with Google anymore
+  # but they will be assigned to this Google group so the user can delete it if they want to.
+  INACTIVE_GROUP_TITLE = 'Inactive'
+
   # Caching the Google contacts from one big request speeds up the sync as we don't need separate HTTP get requests
   # But is only worth it if we are syncing a number of contacts, so check the number against this threshold.
   CACHE_ALL_G_CONTACTS_THRESHOLD = 10
+
+  # The Google sync will wait for any running Google and Facebook imports to finish first to reduce the chance of
+  # unexpected behavior of a sync and an import running at the same time. These constants configure how often to check
+  # for the imports being done and when to just cancel this queuing of the Google contacts sync due to an import
+  # not being finished. (The sync would get re-queued after another contact is saved though).
+  SECS_BETWEEN_CHECK_FOR_IMPORTS_DONE = 30
+  MAX_CHECKS_FOR_IMPORTS_DONE = 60
 
   MAX_CONTACTS_TO_SYNC_CHECKS = 3
 
@@ -18,8 +29,14 @@ class GoogleContactsIntegrator
   end
 
   def sync_contacts
-    # This sync_contacts method is queued when a user in MPDX modifies a contact. The contacts_to_sync will query
-    # the database for contacts that have been modified since they were last synced.
+    # This sync_contacts method is queued when a user in MPDX modifies a contact.
+
+    # That means the imports will also queue the Google Contacts sync, but don't run the sync while an import is
+    # going (especially an import from Google) to allow the import and sync to be more efficient and have less chance
+    # of unexpected interactions
+    return if wait_for_imports_timed_out
+
+    # The contacts_to_sync will query the database for contacts that have been modified since they were last synced.
     #
     # However, it's possible that after editing on contact, the user will then edit another contact before the sync has
     # finished. In that case, sidekiq would attempt to queue the contact sync job, but it would not queue the job because
@@ -35,16 +52,29 @@ class GoogleContactsIntegrator
     # synced after the next contact was modified).
     #
     # Because due to unforseen circumstances, I could imagine that turning into a wasteful infinite loop, I cut off
-    # the number of checks at a constant value of 10.
+    # the number of checks at a constant value.
     contacts_to_sync_checks = 0
     until sync_and_return_num_synced == 0 || contacts_to_sync_checks >= MAX_CONTACTS_TO_SYNC_CHECKS
       contacts_to_sync_checks += 1
     end
+
+    cleanup_inactive_g_contacts
   rescue Person::GoogleAccount::MissingRefreshToken
     # Don't log this exception as we expect it to happen from time to time.
     # Person::GoogleAccount will turn off the contacts integration and send the user an email to refresh their Google login.
   rescue => e
     Airbrake.raise_or_notify(e)
+  end
+
+  def wait_for_imports_timed_out
+    checks_for_imports_done = 0
+    loop do
+      return false if @integration.account_list.imports.where(importing: true).empty?
+      sleep(SECS_BETWEEN_CHECK_FOR_IMPORTS_DONE)
+      checks_for_imports_done += 1
+      break if checks_for_imports_done >= MAX_CHECKS_FOR_IMPORTS_DONE
+    end
+    true
   end
 
   def sync_and_return_num_synced
@@ -72,7 +102,6 @@ class GoogleContactsIntegrator
     # despite re-applying the sync logic.
     @contacts_to_retry_sync.each(&:reload)
     @contacts_to_retry_sync.each(&method(:sync_contact))
-    api_user.send_batched_requests
 
     delete_g_contact_merge_losers
 
@@ -82,22 +111,68 @@ class GoogleContactsIntegrator
   end
 
   def delete_g_contact_merge_losers
-    g_contact_merge_losers.each do |g_contact_link|
-      api_user.delete_contact(g_contact_link.remote_id, g_contact_link.last_etag)
-      g_contact_link.destroy
-    end
-  rescue => e
-    Airbrake.raise_or_notify(e)
+    g_contact_merge_losers.each(&method(:delete_g_contact_merge_loser))
   end
 
   def g_contact_merge_losers
-    # Return the google_contacts record for this Google account that are no longer associated with a contact-person pair
+    # Return the google_contacts for this Google account that are no longer associated with a contact-person pair
     # due to that contact or person being merged with another contact or person in MPDX.
+    # Look for records that have both the contact_id and person_id set, because the Google Import uses an old method
+    #of just setting the person_id and we don't want to wrongly interpret imported Google contacts as merge losers.
     @account.google_contacts
       .joins('LEFT JOIN contact_people ON '\
         'google_contacts.person_id = contact_people.person_id AND contact_people.contact_id = google_contacts.contact_id')
-      .where('contact_people.id IS NULL')
-      .readonly(false)
+      .where('contact_people.id IS NULL').where.not(contact_id: nil).where.not(person_id: nil).readonly(false)
+  end
+
+  def delete_g_contact_merge_loser(g_contact_link)
+    api_user.delete_contact(g_contact_link.remote_id)
+    g_contact_link.destroy
+  rescue => e
+    if defined?(e.response) && GoogleContactsApi::Api.parse_response_code(e.response) == 404
+      # If a 404 error occurred it means the remote Google contact was already deleted, so just delete the link
+      g_contact_link.destroy
+    else
+      # Otherwise we couldn't successfully delete the remote Google Contact, so keep the link to try deleting again later
+      Airbrake.raise_or_notify(e)
+    end
+  end
+
+  def cleanup_inactive_g_contacts
+    GoogleContact.joins(:contact).where(contacts: { account_list_id: @integration.account_list.id })
+      .where(Contact.inactive_conditions).where(google_account: @account).readonly(false)
+      .each(&method(:cleanup_inactive_g_contact))
+    api_user.send_batched_requests
+  end
+
+  def cleanup_inactive_g_contact(g_contact_link, num_retries = 1)
+    g_contact = @cache.find_by_id(g_contact_link.remote_id)
+    unless g_contact
+      g_contact_link.destroy
+      return
+    end
+
+    g_contact.prep_add_to_group(inactive_group)
+    api_user.batch_create_or_update(g_contact) do |status|
+      inactive_cleanup_response(status, g_contact, g_contact_link, num_retries)
+    end
+  end
+
+  def inactive_cleanup_response(status, g_contact, g_contact_link, num_retries)
+    case status[:code]
+    when 200, 404
+      # For success or contact not found, just go ahead and delete the link
+      g_contact_link.destroy
+    when 412
+      fail status.inspect if num_retries < 1
+      # For 412 Etags Mismatch, remove the cached Google contact and retry the operation again
+      @cache.remove_g_contact(g_contact)
+      cleanup_inactive_g_contact(g_contact_link, num_retries - 1)
+    else
+      fail status.inspect
+    end
+  rescue => e
+    Airbrake.raise_or_notify(e)
   end
 
   def api_user
@@ -112,7 +187,7 @@ class GoogleContactsIntegrator
 
   def contacts_to_sync
     if @integration.contacts_last_synced
-      updated_g_contacts = api_user.contacts_updated_min(@integration.contacts_last_synced)
+      updated_g_contacts = api_user.contacts_updated_min(@integration.contacts_last_synced, showdeleted: false)
 
       queried_contacts_to_sync = contacts_to_sync_query(updated_g_contacts)
       if queried_contacts_to_sync.length > CACHE_ALL_G_CONTACTS_THRESHOLD
@@ -135,38 +210,44 @@ class GoogleContactsIntegrator
   # - Or contacts whose google_contacts records have been updated remotely since the last sync as specified by the
   #   updated_g_contacts list
   def contacts_to_sync_query(updated_g_contacts)
+    updated_link_ids = updated_g_contact_links(updated_g_contacts).map(&:id)
+
     @integration.account_list.active_contacts
       .joins(:people)
       .joins("LEFT JOIN addresses ON addresses.addressable_id = contacts.id AND addresses.addressable_type = 'Contact'")
       .joins('LEFT JOIN email_addresses ON people.id = email_addresses.person_id')
       .joins('LEFT JOIN phone_numbers ON people.id = phone_numbers.person_id')
       .joins('LEFT JOIN person_websites ON people.id = person_websites.person_id')
-      .joins('LEFT JOIN google_contacts ON google_contacts.person_id = people.id')
-      .joins("LEFT JOIN #{ remote_g_contacts_sql(updated_g_contacts) } ON remote_g_contacts.remote_id = google_contacts.remote_id")
-      .where('google_contacts.id IS NULL OR google_contacts.google_account_id = ?', @account.id)
-      .group('contacts.id, google_contacts.last_synced, google_contacts.remote_id')
+      .joins('LEFT JOIN google_contacts ON google_contacts.person_id = people.id AND google_contacts.contact_id = contacts.id '\
+          "AND google_contacts.google_account_id = #{ quote_sql(@account.id) }")
+      .group('contacts.id, google_contacts.last_synced, google_contacts.id')
       .having('google_contacts.last_synced IS NULL ' \
         'OR google_contacts.last_synced < ' \
-          'GREATEST(contacts.updated_at, MAX(remote_g_contacts.updated_at), ' \
-                  'MAX(contact_people.updated_at), MAX(people.updated_at), MAX(addresses.updated_at), ' \
-                  'MAX(email_addresses.updated_at), MAX(phone_numbers.updated_at), MAX(person_websites.updated_at))')
+            'GREATEST(contacts.updated_at, MAX(contact_people.updated_at), MAX(people.updated_at), MAX(addresses.updated_at), ' \
+                'MAX(email_addresses.updated_at), MAX(phone_numbers.updated_at), MAX(person_websites.updated_at))' +
+        (updated_link_ids.empty? ? '' : " OR google_contacts.id IN (#{ quote_sql_list(updated_link_ids) })"))
       .distinct
       .readonly(false)
   end
 
-  # Create a SELECT statement to represent the (remote_id, updated) pairs for remotely updated Google Contacts
-  # so that we can join to them via the remote_id and then compare with the updated in the query above.
-  # It's helpful to compare each remote Google Contact's updated field rather than just assume all that are updated
-  # since the overall last sync time, because we calculate the last sync time from the start of the sync to catch
-  # any Google Contacts that were updated during the sync. But the sync itself causes updated contacts since the
-  # start of it, so comparing each one will eliminate double syncing those that were changed by the sync itself,
-  # but will allow capture of remotely changed Google Contacts that were changed by the user during the sync.
-  def remote_g_contacts_sql(g_contacts)
-    sql_table = '(SELECT CAST(null as varchar) as remote_id, CAST(null as timestamp) as updated_at'
-    g_contacts.each do |g_contact|
-      sql_table += " UNION SELECT #{ quote_sql(g_contact.id) }, #{ quote_sql(g_contact.updated.to_s(:db)) }"
+  # Finds Google contact link record that have been remotely modified since their last sync given a list of recently
+  # modified remote Google contacts
+  def updated_g_contact_links(updated_g_contacts)
+    updated_remote_ids = updated_g_contacts.map(&:id)
+    return [] if updated_remote_ids.empty? # Need to check after map as updated_g_contacts is GoogleContactsApi::ResultSet
+
+    g_contacts_by_id = Hash[updated_g_contacts.map { |g_contact| [g_contact.id, g_contact] }]
+    g_contact_links = GoogleContact.select(:id, :remote_id, :last_synced)
+      .where(google_account: @account).where(remote_id: updated_remote_ids).where.not(last_synced: nil).to_a
+
+    g_contact_links.select do |g_contact_link|
+      g_contact = g_contacts_by_id[g_contact_link.remote_id]
+      g_contact.updated > g_contact_link.last_synced
     end
-    sql_table + ') as remote_g_contacts'
+  end
+
+  def quote_sql_list(list)
+    list.map { |item| ActiveRecord::Base.connection.quote(item) }.join(',')
   end
 
   def quote_sql(sql)
@@ -205,13 +286,25 @@ class GoogleContactsIntegrator
     [g_contact, g_contact_link]
   end
 
+  def groups
+    @groups ||= api_user.groups
+  end
+
+  def inactive_group
+    @inactive_group ||= groups.find { |group| group.title == INACTIVE_GROUP_TITLE } ||
+      GoogleContactsApi::Group.create({ title: INACTIVE_GROUP_TITLE }, api_user.api)
+  end
+
   def mpdx_group
-    @mpdx_group ||= api_user.groups.find { |group| group.title == CONTACTS_GROUP_TITLE } ||
+    @mpdx_group ||= groups.find { |group| group.title == CONTACTS_GROUP_TITLE } ||
       GoogleContactsApi::Group.create({ title: CONTACTS_GROUP_TITLE }, api_user.api)
   end
 
   def get_or_query_g_contact(g_contact_link, person)
-    return @cache.find_by_id(g_contact_link.remote_id) if g_contact_link.remote_id
+    if g_contact_link.remote_id
+      retrieved_g_contact = @cache.find_by_id(g_contact_link.remote_id)
+      return retrieved_g_contact if retrieved_g_contact
+    end
 
     @cache.select_by_name(person.first_name, person.last_name).find do |g_contact|
       !@assigned_remote_ids.include?(g_contact.id)
